refactor page-service
