+++ src/entities/common-interfaces.ts
@@ -45,18 +45,3 @@ export interface AbleToExportAndImportEuring {
   exportEURING(): string;
   importEURING(code: string): any;
 }
-
-export enum SortingDirection {
-  asc = 'ASC',
-  desk = 'DESC',
-}
-
-export interface Search {
-  search?: string;
-  pageNumber?: number;
-  pageSize?: number;
-  sortingColumn?: string;
-  sortingDirection?: SortingDirection;
-}
-
-export type Locale = 'eng' | 'rus' | 'byn';
