+++ src/services/export/XLSExporterForObservations.ts
@@ -3,7 +3,7 @@ import { write, utils } from 'xlsx';
 import AbstractExporter, { ExporterType } from './AbstractExporter';
 import { Observation } from '../../entities/observation-entity';
 import { User } from '../../entities/user-entity';
-import { Locale } from '../../entities/common-interfaces';
+import { Locale } from '../page-service';
 
 export default class XLSExporterForObservations extends AbstractExporter<Buffer> {
   public type: ExporterType = ExporterType.xls;
@@ -58,12 +58,12 @@ export default class XLSExporterForObservations extends AbstractExporter<Buffer>
       relations: this.ObservationColumnsByDesc,
     });
 
-      // sanitize user's sensetive data
-      observations.map(obs => {
-        const ref = obs;
-        ref.finder = User.sanitizeUser(ref.finder) as User;
-        return ref;
-      });
+    // sanitize user's sensetive data
+    observations.map(obs => {
+      const ref = obs;
+      ref.finder = User.sanitizeUser(ref.finder) as User;
+      return ref;
+    });
 
     const flattenObservations = observations.map(obs => this.flattenObservation(obs));
     const workSheet = utils.json_to_sheet(flattenObservations);
